refactor get_port_mode
